backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud services
google-cloud-bigquery==3.13.0
google-cloud-storage==2.10.0
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Data processing
pandas==2.1.4
numpy==1.25.2
python-dateutil==2.8.2

# CSV and file handling
python-multipart==0.0.6
aiofiles==23.2.1

# HTTP client for scraper communication
httpx==0.25.2
aiohttp==3.9.1

# Logging and monitoring
loguru==0.7.2
structlog==23.2.0

# Environment and configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Database and async support
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Background tasks
celery==5.3.4
redis==5.0.1

# API documentation
python-multipart==0.0.6
